Refactor product request models and update dependency management

+++ src/main/java/com/mertalptekin/springrestapplication/application/request/product/ProductUpdateRequest.java
@@ -2,9 +2,11 @@ package com.mertalptekin.springrestapplication.application.request.product;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 // Hangi ürünün hangi alanlarını güncelleyeceğiz ?
 public record ProductUpdateRequest(@JsonProperty("product_id") Integer id,
                                    @JsonProperty("product_name") String name,
-                                   @JsonProperty("product_price") Double price,
+                                   @JsonProperty("product_price") BigDecimal price,
                                    @JsonProperty("product_quantity") Integer quantity) {
 }
